feat: Add json customized marshaler for uint64
